DuringWorkoutViewModel 생성 및 데이터 연결

+++ WorkoutDone/WorkoutDone/Sources/Model/WorkOutDone/WeightTrainingInfo.swift
@@ -10,10 +10,10 @@ import RealmSwift
 
 class WeightTrainingInfo : Object {
     @Persisted dynamic var setCount : Int
-    @Persisted dynamic var weight : Int?
+    @Persisted dynamic var weight : Double?
     @Persisted dynamic var trainingCount : Int?
     
-    convenience init(setCount: Int, weight: Int, trainingCount: Int) {
+    convenience init(setCount: Int, weight: Double, trainingCount: Int) {
         self.init()
         self.setCount = setCount
         self.weight = weight
